refactor(27085): refactor the toolbar
